+++ build.gradle
@@ -1,19 +1,17 @@
 buildscript {
     ext {
-        agpVersion = "4.2.2"
-        kotlinVersion = "1.5.30"
-        ktxCoreVersion = "1.6.0"
-        appCompatVersion = "1.3.1"
-        coroutinesVersion = "1.5.2"
-        retrofitVersion = "2.9.0"
-        moshiVersion = "1.12.0"
-        moshiRetrofitVersion = "2.9.0"
-        simpleXmlVersion = "2.9.0"
+        agpVersion = "8.2.0"
+        kotlinVersion = "1.9.23"
+        coroutinesVersion = '1.8.0'
+        retrofitVersion = '2.10.0'
+        moshiVersion = '1.15.1'
+        moshiRetrofitVersion = '2.10.0'
+        simpleXmlVersion = '2.10.0'
         ktLintVersion = "10.1.0"
-        coroutineCallAdapterVersion = "0.9.2"
         jUnitVersion = "4.13.2"
-        mockkVersion = "1.12.0"
+        mockkVersion = '1.13.10'
     }
+
     repositories {
         google()
         mavenCentral()
@@ -21,14 +19,16 @@ buildscript {
             url "https://plugins.gradle.org/m2/"
         }
     }
+
     dependencies {
         classpath "com.android.tools.build:gradle:$agpVersion"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
         classpath "org.jlleitschuh.gradle:ktlint-gradle:$ktLintVersion"
     }
 }
+
 plugins {
-    id 'com.google.devtools.ksp' version '1.8.10-1.0.9' apply false
+    id 'com.google.devtools.ksp' version '1.9.23-1.0.19' apply false
 }
 
 allprojects {
