+++ library/src/main/java/com/sharkaboi/appupdatechecker/provider/AppUpdateServices.kt
@@ -1,6 +1,5 @@
 package com.sharkaboi.appupdatechecker.provider
 
-import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
 import com.sharkaboi.appupdatechecker.sources.fdroid.FdroidConstants
 import com.sharkaboi.appupdatechecker.sources.fdroid.FdroidService
 import com.sharkaboi.appupdatechecker.sources.github.GithubConstants
@@ -18,7 +17,6 @@ internal object AppUpdateServices {
     val fDroidService: FdroidService by lazy(LazyThreadSafetyMode.SYNCHRONIZED) {
         Retrofit.Builder()
             .baseUrl(FdroidConstants.BASE_URL)
-            .addCallAdapterFactory(CoroutineCallAdapterFactory())
             .addConverterFactory(MoshiConverterFactory.create())
             .build()
             .create(FdroidService::class.java)
@@ -27,7 +25,6 @@ internal object AppUpdateServices {
     val jsonService: JsonService by lazy(LazyThreadSafetyMode.SYNCHRONIZED) {
         Retrofit.Builder()
             .baseUrl(JsonConstants.BASE_URL)
-            .addCallAdapterFactory(CoroutineCallAdapterFactory())
             .addConverterFactory(MoshiConverterFactory.create())
             .build()
             .create(JsonService::class.java)
@@ -36,7 +33,6 @@ internal object AppUpdateServices {
     val xmlService: XMLService by lazy(LazyThreadSafetyMode.SYNCHRONIZED) {
         Retrofit.Builder()
             .baseUrl(XMLConstants.BASE_URL)
-            .addCallAdapterFactory(CoroutineCallAdapterFactory())
             .addConverterFactory(SimpleXmlConverterFactory.create())
             .build()
             .create(XMLService::class.java)
@@ -45,7 +41,6 @@ internal object AppUpdateServices {
     val githubService: GithubService by lazy(LazyThreadSafetyMode.SYNCHRONIZED) {
         Retrofit.Builder()
             .baseUrl(GithubConstants.BASE_URL)
-            .addCallAdapterFactory(CoroutineCallAdapterFactory())
             .addConverterFactory(MoshiConverterFactory.create())
             .build()
             .create(GithubService::class.java)
