+++ library/src/main/java/com/sharkaboi/appupdatechecker/sources/fdroid/FdroidResponse.kt
@@ -1,10 +1,8 @@
 package com.sharkaboi.appupdatechecker.sources.fdroid
 
-import androidx.annotation.Keep
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
-@Keep
 @JsonClass(generateAdapter = true)
 data class FdroidResponse(
     @Json(name = "packageName")
@@ -12,7 +10,6 @@ data class FdroidResponse(
     @Json(name = "packages")
     val packages: List<Package>
 ) {
-    @Keep
     @JsonClass(generateAdapter = true)
     data class Package(
         @Json(name = "versionCode")
