+++ library/src/main/java/com/sharkaboi/appupdatechecker/sources/fdroid/FdroidService.kt
@@ -7,7 +7,7 @@ import retrofit2.http.Path
 internal interface FdroidService {
 
     @GET(FdroidConstants.PATH)
-    fun getReleasesAsync(
+    suspend fun getReleasesAsync(
         @Path(FdroidConstants.PACKAGE_PATH_ID) packageName: String
-    ): Deferred<FdroidResponse>
+    ): FdroidResponse
 }
