+++ library/src/main/java/com/sharkaboi/appupdatechecker/sources/json/JsonResponse.kt
@@ -1,10 +1,8 @@
 package com.sharkaboi.appupdatechecker.sources.json
 
-import androidx.annotation.Keep
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
-@Keep
 @JsonClass(generateAdapter = true)
 data class JsonResponse(
     @Json(name = "latestVersion")
