+++ library/src/test/java/com/github/sharkaboi/appupdatechecker/GithubTest.kt
@@ -1,6 +1,7 @@
-package com.sharkaboi.appupdatechecker
+package com.github.sharkaboi.appupdatechecker
 
 import android.content.Context
+import com.sharkaboi.appupdatechecker.AppUpdateChecker
 import com.sharkaboi.appupdatechecker.extensions.isInternetConnected
 import com.sharkaboi.appupdatechecker.models.AppUpdateCheckerSource
 import com.sharkaboi.appupdatechecker.models.UpdateState
@@ -11,7 +12,7 @@ import kotlinx.coroutines.runBlocking
 import org.junit.Before
 import org.junit.Test
 
-class XMLTest {
+class GithubTest {
     private lateinit var context: Context
 
     @Before
@@ -25,8 +26,9 @@ class XMLTest {
     fun `Checker on older installed version returns new version`() = runBlocking {
         val testChecker = AppUpdateChecker(
             context,
-            source = AppUpdateCheckerSource.XMLSource(
-                xmlEndpoint = "https://gist.githubusercontent.com/Sharkaboi/66b45a22afde23a9b2781eeec6f10c56/raw/132296a0e85d254cd982959bc8f198ca61c213b7/test.xml"
+            source = AppUpdateCheckerSource.GithubSource(
+                ownerUsername = "Sharkaboi",
+                repoName = "MediaHub"
             ),
             currentVersionTag = "v0.0.0"
         )
@@ -39,8 +41,9 @@ class XMLTest {
     fun `Checker on newer installed version returns no update`() = runBlocking {
         val testChecker = AppUpdateChecker(
             context,
-            source = AppUpdateCheckerSource.XMLSource(
-                xmlEndpoint = "https://gist.githubusercontent.com/Sharkaboi/66b45a22afde23a9b2781eeec6f10c56/raw/132296a0e85d254cd982959bc8f198ca61c213b7/test.xml"
+            source = AppUpdateCheckerSource.GithubSource(
+                ownerUsername = "Sharkaboi",
+                repoName = "MediaHub"
             ),
             currentVersionTag = "v${Int.MAX_VALUE}.0.0"
         )
@@ -50,58 +53,77 @@ class XMLTest {
     }
 
     @Test
-    fun `Checker on invalid xml repo returns invalid error`() = runBlocking {
+    fun `Checker on invalid github repo returns invalid error`() = runBlocking {
         val testChecker = AppUpdateChecker(
             context,
-            source = AppUpdateCheckerSource.XMLSource(
-                xmlEndpoint = "https://google.com"
+            source = AppUpdateCheckerSource.GithubSource(
+                ownerUsername = "Sharkaboi",
+                repoName = "adadadadadadadadaadada"
             ),
             currentVersionTag = "v0.0.0"
         )
         val result = testChecker.checkUpdate()
         println(result)
-        assert(result is UpdateState.XMLInvalid)
+        assert(result is UpdateState.GithubInvalid)
     }
 
     @Test
-    fun `Checker on invalid xml schema returns invalid error`() = runBlocking {
+    fun `Checker on invalid github user returns invalid error`() = runBlocking {
         val testChecker = AppUpdateChecker(
             context,
-            source = AppUpdateCheckerSource.XMLSource(
-                xmlEndpoint = "https://gist.githubusercontent.com/Sharkaboi/66b45a22afde23a9b2781eeec6f10c56/raw/132296a0e85d254cd982959bc8f198ca61c213b7/invalid.xml"
+            source = AppUpdateCheckerSource.GithubSource(
+                ownerUsername = "adadadadadadadadaadada",
+                repoName = "MediaHub"
             ),
             currentVersionTag = "v0.0.0"
         )
         val result = testChecker.checkUpdate()
         println(result)
-        assert(result is UpdateState.XMLInvalid)
+        assert(result is UpdateState.GithubInvalid)
     }
 
     @Test
-    fun `Checker on invalid url returns malformed error`() = runBlocking {
+    fun `Checker on github with no release returns invalid error`() = runBlocking {
         val testChecker = AppUpdateChecker(
             context,
-            source = AppUpdateCheckerSource.XMLSource(
-                xmlEndpoint = "invalid url"
+            source = AppUpdateCheckerSource.GithubSource(
+                ownerUsername = "Sharkaboi",
+                repoName = "sharkaboi.github.io"
             ),
             currentVersionTag = "v0.0.0"
         )
         val result = testChecker.checkUpdate()
         println(result)
-        assert(result is UpdateState.XMLMalformed)
+        assert(result is UpdateState.GithubInvalid)
     }
 
     @Test
-    fun `Checker on blank endpoint returns malformed error`() = runBlocking {
+    fun `Checker on blank username returns malformed error`() = runBlocking {
         val testChecker = AppUpdateChecker(
             context,
-            source = AppUpdateCheckerSource.XMLSource(
-                xmlEndpoint = "   "
+            source = AppUpdateCheckerSource.GithubSource(
+                ownerUsername = "   ",
+                repoName = "sharkaboi.github.io"
             ),
             currentVersionTag = "v0.0.0"
         )
         val result = testChecker.checkUpdate()
         println(result)
-        assert(result is UpdateState.XMLMalformed)
+        assert(result is UpdateState.GithubMalformed)
+    }
+
+    @Test
+    fun `Checker on empty repo name returns malformed error`() = runBlocking {
+        val testChecker = AppUpdateChecker(
+            context,
+            source = AppUpdateCheckerSource.GithubSource(
+                ownerUsername = "Sharkaboi",
+                repoName = ""
+            ),
+            currentVersionTag = "v0.0.0"
+        )
+        val result = testChecker.checkUpdate()
+        println(result)
+        assert(result is UpdateState.GithubMalformed)
     }
 }
